+++ tests/tests_args.rs
@@ -5,10 +5,10 @@ mod tests {
     fn test_parse_args() {
         let args = vec![
             (
-                vec!["detox".to_owned(), "README.md".to_owned()],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: true },
-                    verbosity: detox::VerbosityFields {
+                vec!["filerefine".to_owned(), "README.md".to_owned()],
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: true },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: true,
                         json: false,
                         json_pretty: false,
@@ -17,10 +17,10 @@ mod tests {
                 },
             ),
             (
-                vec!["detox".to_owned(), "README.md".to_owned(), "-d".to_owned()],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: false },
-                    verbosity: detox::VerbosityFields {
+                vec!["filerefine".to_owned(), "README.md".to_owned(), "-d".to_owned()],
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: false },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: true,
                         json: false,
                         json_pretty: false,
@@ -30,14 +30,14 @@ mod tests {
             ),
             (
                 vec![
-                    "detox".to_owned(),
+                    "filerefine".to_owned(),
                     "README.md".to_owned(),
                     "-d".to_owned(),
                     "-j".to_owned(),
                 ],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: false },
-                    verbosity: detox::VerbosityFields {
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: false },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: false,
                         json: true,
                         json_pretty: false,
@@ -47,14 +47,14 @@ mod tests {
             ),
             (
                 vec![
-                    "detox".to_owned(),
+                    "filerefine".to_owned(),
                     "README.md".to_owned(),
                     "-d".to_owned(),
                     "--json".to_owned(),
                 ],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: false },
-                    verbosity: detox::VerbosityFields {
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: false },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: false,
                         json: true,
                         json_pretty: false,
@@ -64,14 +64,14 @@ mod tests {
             ),
             (
                 vec![
-                    "detox".to_owned(),
+                    "filerefine".to_owned(),
                     "README.md".to_owned(),
                     "-d".to_owned(),
                     "-e".to_owned(),
                 ],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: false },
-                    verbosity: detox::VerbosityFields {
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: false },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: false,
                         json: true,
                         json_pretty: false,
@@ -81,14 +81,14 @@ mod tests {
             ),
             (
                 vec![
-                    "detox".to_owned(),
+                    "filerefine".to_owned(),
                     "README.md".to_owned(),
                     "-d".to_owned(),
                     "--json-error".to_owned(),
                 ],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: false },
-                    verbosity: detox::VerbosityFields {
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: false },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: false,
                         json: true,
                         json_pretty: false,
@@ -98,14 +98,14 @@ mod tests {
             ),
             (
                 vec![
-                    "detox".to_owned(),
+                    "filerefine".to_owned(),
                     "README.md".to_owned(),
                     "-d".to_owned(),
                     "-p".to_owned(),
                 ],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: false },
-                    verbosity: detox::VerbosityFields {
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: false },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: false,
                         json: true,
                         json_pretty: true,
@@ -115,14 +115,14 @@ mod tests {
             ),
             (
                 vec![
-                    "detox".to_owned(),
+                    "filerefine".to_owned(),
                     "README.md".to_owned(),
                     "-d".to_owned(),
                     "--json-pretty".to_owned(),
                 ],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: false },
-                    verbosity: detox::VerbosityFields {
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: false },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: false,
                         json: true,
                         json_pretty: true,
@@ -131,10 +131,10 @@ mod tests {
                 },
             ),
             (
-                vec!["detox".to_owned(), "-v".to_owned()],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: true },
-                    verbosity: detox::VerbosityFields {
+                vec!["filerefine".to_owned(), "-v".to_owned()],
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: true },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: false,
                         json: true,
                         json_pretty: true,
@@ -143,10 +143,10 @@ mod tests {
                 },
             ),
             (
-                vec!["detox".to_owned(), "--version".to_owned()],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: true },
-                    verbosity: detox::VerbosityFields {
+                vec!["filerefine".to_owned(), "--version".to_owned()],
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: true },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: false,
                         json: true,
                         json_pretty: true,
@@ -155,10 +155,10 @@ mod tests {
                 },
             ),
             (
-                vec!["detox".to_owned(), "-q".to_owned()],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: true },
-                    verbosity: detox::VerbosityFields {
+                vec!["filerefine".to_owned(), "-q".to_owned()],
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: true },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: false,
                         json: false,
                         json_pretty: false,
@@ -167,10 +167,10 @@ mod tests {
                 },
             ),
             (
-                vec!["detox".to_owned(), "--quiet".to_owned()],
-                detox::OptionnalFields {
-                    options: detox::OptionsFields { dry_run: true },
-                    verbosity: detox::VerbosityFields {
+                vec!["filerefine".to_owned(), "--quiet".to_owned()],
+                filerefine::OptionnalFields {
+                    options: filerefine::OptionsFields { dry_run: true },
+                    verbosity: filerefine::VerbosityFields {
                         verbose: false,
                         json: false,
                         json_pretty: false,
@@ -181,7 +181,7 @@ mod tests {
         ];
         print!("ARGS: {:?}\n", args);
         for one_test in args.iter() {
-            let res = detox::parse_args(one_test.0.clone());
+            let res = filerefine::parse_args(one_test.0.clone());
             if let Ok(ok_res) = res {
                 assert_eq!(ok_res.0, one_test.1);
             } else {
@@ -193,17 +193,17 @@ mod tests {
     #[test]
     fn test_parse_args_no_file_found() {
         let vec_args = vec![
-            "detox".to_owned(),
+            "filerefine".to_owned(),
             "README.md".to_owned(),
             "README".to_owned(),
         ];
-        let res = detox::parse_args(vec_args);
+        let res = filerefine::parse_args(vec_args);
         let (options, vect) = res.ok().unwrap();
         assert_eq!(
             options,
-            detox::OptionnalFields {
-                options: detox::OptionsFields { dry_run: true },
-                verbosity: detox::VerbosityFields {
+            filerefine::OptionnalFields {
+                options: filerefine::OptionsFields { dry_run: true },
+                verbosity: filerefine::VerbosityFields {
                     verbose: true,
                     json: false,
                     json_pretty: false,
@@ -216,14 +216,14 @@ mod tests {
 
     #[test]
     fn test_parse_args_star() {
-        let vec_args = vec!["detox".to_owned(), "*".to_owned()];
-        let res = detox::parse_args(vec_args);
+        let vec_args = vec!["filerefine".to_owned(), "*".to_owned()];
+        let res = filerefine::parse_args(vec_args);
         let (options, vect) = res.ok().unwrap();
         assert_eq!(
             options,
-            detox::OptionnalFields {
-                options: detox::OptionsFields { dry_run: true },
-                verbosity: detox::VerbosityFields {
+            filerefine::OptionnalFields {
+                options: filerefine::OptionsFields { dry_run: true },
+                verbosity: filerefine::VerbosityFields {
                     verbose: true,
                     json: false,
                     json_pretty: false,